Added a supply depo, supply station and a harvester.
TODO: Make sure the harvester gets a target as soon as it spawns.
The harvester can now be built only from a supply station and if you click on it and target a supply depo it will go to it, collect enough money that it can store and then return to the supply station to drop the money off to the player.